Internet connection check, Player animations and animated gradient, general clean up
Note that internet connection check does NOT work perfectly fine as of now. It works but i need to handle timeouts better and tweak the frequency of the connection checks. Controls still not functioning, and I'm considering the option to make the Control's view transparent as well so the gradient background can be seen. Also, i need to tweak the disk rotation animation